fix: add micromark parser to markdownlint config
In markdownlint 0.36.0, the support for micromark returned. Adding it in the config
object fixes the error described in
[this issue](https://github.com/silvermine/markdownlint-rule-indent-alignment/issues/7)

The new micromark version changes some previously assumed parser behaviors. It causes
some errors related to nested intentation in HTML snippets. Since currently there is no
need to lint HTML, the tokens related to it are now ignored.

Introduces a new test verification to prevent an error where  tokens cause
other text tokens to appear to be indented.